+++ test/containers/classy/int_test.exs
@@ -1,0 +1,4 @@
+defmodule Containers.Classy.IntegerTest do
+  use ExUnit.Case, async: true
+  doctest Containers.Classy.Integer
+end