+++ test/containers/int_text.exs
@@ -1,4 +1,0 @@
-defmodule Containers.IntTest do
-  use ExUnit.Case, async: true
-  doctest Containers.Int
-end